Modify fetch mission action
